Move tests into separate files
