+++ package.json
@@ -4,7 +4,7 @@
   "description": "Diacritic transliteration tools using diacritics.io API",
   "main": "index.js",
   "scripts": {
-    "test": "ava",
+    "test": "ava test/*.js",
     "linting": "./node_modules/.bin/eslint './*.js'"
   },
   "repository": {
