+++ test/regexp.js
@@ -1,167 +1,11 @@
 import test from "ava";
-import d from "./";
-import request from "./node_modules/sync-request";
+import d from "../index.js";
 
-// change boolean to debug this script
-d.debug = {
-    server: false, // log server & cache interactions
-    regexp: false, // log module RegExp creation
-    regexpTests: false // log RegExp tests
-};
-
-// get current diacritics.json directly
-const dist = "https://raw.githubusercontent.com/diacritics/database/dist/v" +
-    d.maxVersion + "/diacritics.json",
-    response = request("GET", dist),
-    diacritics = response.statusCode === 200 ?
-        JSON.parse(response.getBody().toString()) :
-        {};
-
-test("Set & Get API version", t => {
-    const max = d.maxVersion;
-    // bump max version for testing
-    d.maxVersion = 6;
-    t.is(d.setVersion("2"), "v2");
-    t.is(d.setVersion(3.4), "v3");
-    t.is(d.setVersion("x4"), "v4");
-    t.is(d.setVersion("5.675"), "v5");
-    t.is(d.setVersion("x"), "v5");
-    t.is(d.setVersion("v0.1"), "v5");
-    t.is(d.setVersion("9999"), "v5");
-    t.is(d.setVersion(null), "v5");
-    t.is(d.setVersion(undefined), "v5");
-    t.is(d.getVersion(), "v5");
-    t.is(d.setVersion("v1"), "v1");
-    d.maxVersion = max;
-});
-
-test("Get Language", t => {
-    t.deepEqual(d.getLanguage("German"), diacritics.de);
-    t.deepEqual(d.getLanguage("test"), {
-        message: "Language 'test' was not found"
-    });
-    // make sure user set version is handled
-    d.currentVersion = "v999";
-    const error = d.getLanguage("test");
-    d.currentVersion = "v" + d.maxVersion;
-    t.truthy(error.message.match("error: 404"));
-});
-
-test("Get Variant", t => {
-    // ensure current version is set, as all ava tests are run in parallel
-    d.currentVersion = "v" + d.maxVersion;
-    t.deepEqual(d.getVariant("de"), diacritics.de);
-    t.deepEqual(d.getVariant("test"), {
-        message: "Variant 'test' was not found"
-    });
-});
-
-test("Get Continent", t => {
-    // ensure current version is set, as all ava tests are run in parallel
-    d.currentVersion = "v" + d.maxVersion;
-    // Antarctica... this may change once more data is added
-    t.deepEqual(d.getContinent("an"), diacritics.es);
-    t.deepEqual(d.getContinent("test"), {
-        message: "Continent 'test' was not found"
-    });
-});
-
-test("Get Alphabet", t => {
-    // ensure current version is set, as all ava tests are run in parallel
-    d.currentVersion = "v" + d.maxVersion;
-    // only compare one variant, or things could get messy once the database
-    // gets bigger
-    t.deepEqual(d.getAlphabet("latn").es, diacritics.es.es);
-    t.deepEqual(d.getAlphabet("test"), {
-        message: "Alphabet 'test' was not found"
-    });
-});
-
-test("Format unicode", t => {
-    t.is(d.formatUnicode("T\\u00E9st"), "Tést");
-    t.is(d.formatUnicode("\\u00C4 Te\\u0301st"), "Ä Tést");
-});
-
-test("Get Diacritics", t => {
-    // ensure current version is set, as all ava tests are run in parallel
-    d.currentVersion = "v" + d.maxVersion;
-    // only compare a unique diacritics, or things could get messy once the
-    // database gets bigger
-    const result = {
-        es: {
-            metadata: diacritics.es.es.metadata,
-            data: {
-                "ñ": diacritics.es.es.data["ñ"],
-                "¿": diacritics.es.es.data["¿"]
-            }
-        },
-        de: {
-            metadata: diacritics.de.de.metadata,
-            data: {
-                "ß": diacritics.de.de.data["ß"]
-            }
-        }
-    };
-    t.deepEqual(d.getDiacritics("¿abcñ-ß123"), result);
-    t.deepEqual(d.getDiacritics("test"), {
-        message: "No diacritics found"
-    });
-});
-
-test("Get Base", t => {
-    // ensure current version is set, as all ava tests are run in parallel
-    d.currentVersion = "v" + d.maxVersion;
-    // only compare a unique base, or things could get messy once the database
-    // gets bigger
-    const result = {
-        de: {
-            metadata: diacritics.de.de.metadata,
-            data: {
-                "ß": diacritics.de.de.data["ß"]
-            }
-        }
-    };
-    // getBase will accept a single base string, or array
-    t.deepEqual(d.getBase("ß"), result);
-    t.deepEqual(d.getBase(["&"]), {
-        message: "No matching bases found"
-    });
-});
-
-test("Get Decompose", t => {
-    // ensure current version is set, as all ava tests are run in parallel
-    d.currentVersion = "v" + d.maxVersion;
-    // only compare a unique base, or things could get messy once the database
-    // gets bigger
-    const result = {
-        de: {
-            metadata: diacritics.de.de.metadata,
-            data: {
-                "ß": diacritics.de.de.data["ß"]
-            }
-        }
-    };
-    // getBase will accept a single base string, or array
-    t.deepEqual(d.getDecompose("ss"), result);
-    t.deepEqual(d.getDecompose(["&", "test"]), {
-        message: "No matching decomposes found"
-    });
-});
-
-test("Transliterate", t => {
-    // ensure current version is set, as all ava tests are run in parallel
-    d.currentVersion = "v" + d.maxVersion;
-    t.is(d.transliterate("¿abcñ-ß123?", "decompose", "de"), "¿abcñ-ss123?");
-    // base of ß is still ß
-    t.is(d.transliterate("¿abcñ-ß123?", "base", "de"), "¿abcñ-ß123?");
-    t.is(d.transliterate("¿abcñ-ß123?", "decompose", "es"), "?abcn-ß123?");
-    t.is(d.transliterate("¿abcñ-ß123?", "base", "es"), "?abcn-ß123?");
-    t.is(d.transliterate("¿abcñ-ß123?", "decompose"), "?abcn-ss123?");
-    // the next two tests have equivalent results
-    t.is(d.transliterate("¿abcñ-ß123?", "base"), "?abcn-ß123?");
-    t.is(d.transliterate("¿abcñ-ß123?"), "?abcn-ß123?");
-    // no variant found, return original string
-    t.is(d.transliterate("¿abcñ-ß123?", "base", "test"), "¿abcñ-ß123?");
+require("fs").readFile("test/settings.json", "utf8", (err, data) => {
+    if(err) {
+        throw err;
+    }
+    d.debug = JSON.parse(data);
 });
 
 const regexpTests = [{
@@ -347,7 +191,7 @@ const regexpTests = [{
     // finalize callback
     regexp: "tést",
     options: {
-        finalize: function(array, joiner) {
+        done: function(array, joiner) {
             // match whole words only
             array.unshift("\\b");
             array.push("\\b");
@@ -355,17 +199,15 @@ const regexpTests = [{
         }
     },
     tests: [
-        [ "true", "t\u00E9st", "finalize; match whole word" ],
-        [ "true", "te\u0301st", "finalize; match whole word" ],
-        [ "false", "ste\u0301st", "finalize; no internal match" ],
-        [ "false", "t\u00E9sts", "finalize; no internal match" ],
+        [ "true", "t\u00E9st", "done; match whole word" ],
+        [ "true", "te\u0301st", "done; match whole word" ],
+        [ "false", "ste\u0301st", "done; no internal match" ],
+        [ "false", "t\u00E9sts", "done; no internal match" ],
         [ "false", "1234", "defaults; don't match everything"]
     ]
 }];
 
 test("Create RegExp", t => {
-    // ensure current version is set, as all ava tests are run in parallel
-    d.currentVersion = "v" + d.maxVersion;
     let regexp;
     regexpTests.forEach(tst => {
         regexp = d.createRegExp(tst.regexp, tst.options);
@@ -390,12 +232,13 @@ test("Create RegExp", t => {
     });
 });
 
-// used for debugging
+// functions used for debugging
 function debugFormat(val) {
     // return ignore leading zero if hex <= FFFF
     let size = val < 0x10000 ? -4 : -5;
     return "\\u" + ("00000" + val.toString(16)).toUpperCase().slice(size);
 };
+
 function decodeUnicode(str) {
     const len = str.length;
     let chr, low,
