+++ test/settings.json
@@ -1,0 +1,6 @@
+{
+    "server": false,
+    "regexp": false,
+    "regexpTests": false,
+    "placeholder": false
+}
